♻️ R6RZ93 task: split task lifecycle hotspots
